MAINT: split util package to scripts
