+++ dtoolkit/transformer/factory.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from dtoolkit.transformer.base import MethodTF
-from dtoolkit.util.generic import snake_to_camel
+from dtoolkit.util import snake_to_camel
 
 if TYPE_CHECKING:
     from typing import Callable
