+++ dtoolkit/util/snake_to_camel.py
@@ -1,0 +1,24 @@
+def snake_to_camel(name: str) -> str:
+    """
+    Change snake style name to camel style name.
+
+    Parameters
+    ----------
+    name : str
+        Snake style name
+
+    Returns
+    -------
+    str
+        Camel style name
+
+    Examples
+    --------
+    >>> from dtoolkit.util import snake_to_camel
+    >>> snake_to_camel(snake_to_camel.__name__)
+    'SnakeToCamel'
+    """
+
+    components = name.split("_")
+    components = (x.title() for x in components)
+    return "".join(components)
